fix(finish-pr-body): resolve spec-review findings
- _prBodyDeps.readText treats ENOENT as absent (mirrors _qualityDeps),
  instead of warning on the routine missing-artifact case
- fix pr-body.test.ts's FinishRound import (was importing from a module
  that never exported it; masked at runtime by type-only erasure)
- move pr-body.ts into flows/nax-finish/steps/pr-body.ts and fold
  loadFinishPrContext into it, matching the spec's stated module
  boundary (seam S3); steps/pr.ts re-exports _prDeps/loadFinishPrContext
  for backward compatibility
